Fix spacing in Task.java
Previously, when a task did not have a tag, there was still a trailing
space at the end of the various tasks.toString methods.

Although this cant be seen on the output, it fails previous test cases.

I change the task.getTag() method to only include the space if there is
a tag present.

This way, the various task.toString() methods will remain the same if a
task does not have a tag.